float math

+++ src/Adsr.cpp
@@ -50,8 +50,8 @@ void Adsr::configureTrig(Signal *trig) {
    _trig = trig;
 }
 
-double IRAM_ATTR Adsr::update(uint64_t time) {
-   double output;
+float IRAM_ATTR Adsr::update(uint64_t time) {
+   float output;
 
    bool trigIn = _trigThreshold->compute(_trig->read(time));
    bool gateIn = _gateThreshold->compute(_gate->read(time));
@@ -68,9 +68,9 @@ double IRAM_ATTR Adsr::update(uint64_t time) {
       switch (stateMachine) {
          case 1:  // Attack
          {
-            double attackIn = _attack->read(time);
-            double k = 20.0 / slopeTime(attackIn);
-            double m = -10.0;
+            float attackIn = _attack->read(time);
+            float k = 20.0 / slopeTime(attackIn);
+            float m = -10.0;
             output = (time - startedAt) * k + m;
             if (output >= 10.0) {
                startedAt = time;
@@ -80,10 +80,10 @@ double IRAM_ATTR Adsr::update(uint64_t time) {
          }
          case 2:  // Decay
          {
-            double sustainIn = _sustain->read(time);
-            double decayIn = _decay->read(time);
-            double k = -(10 - sustainIn) / slopeTime(decayIn);
-            double m = 10;
+            float sustainIn = _sustain->read(time);
+            float decayIn = _decay->read(time);
+            float k = -(10 - sustainIn) / slopeTime(decayIn);
+            float m = 10;
             output = (time - startedAt) * k + m;
             if (output >= 10.0) {
                startedAt = time;
@@ -100,10 +100,10 @@ double IRAM_ATTR Adsr::update(uint64_t time) {
             output = -10;
       }
    } else {
-      double releaseIn = _release->read(time);
-      double sustainIn = _sustain->read(time);
-      double k = -sustainIn / slopeTime(releaseIn);
-      double m = sustainIn;
+      float releaseIn = _release->read(time);
+      float sustainIn = _sustain->read(time);
+      float k = -sustainIn / slopeTime(releaseIn);
+      float m = sustainIn;
       output = (time - startedAt) * k + m;
       if (gateIn) {
          // Start new cycle
@@ -115,10 +115,10 @@ double IRAM_ATTR Adsr::update(uint64_t time) {
    return Demiurge::clip(output);
 }
 
-double Adsr::slopeTime(double voltage) {
+float Adsr::slopeTime(float voltage) {
    // Logarithmic response, so that;
    // -10V = 1 microsecond, 0V = 1 millisecond, +10V = 1 second
    // TODO: If log() is too slow, use lookup tables.
-   double millis = pow10(voltage / 3.33333);
+   float millis = pow10(voltage / 3.33333);
    return millis * 1000;
 }
